manager/pyproject.toml
[project]
name = "manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "moss-voice-agent-manager>=1.0.0b3",
    "python-dotenv>=1.2.1",
]
